requirements.txt
Flask==2.1.0; python_version > '3.6'
Flask==2.0.3; python_version < '3.7'
gunicorn==20.1.0
google-cloud-error-reporting==1.8.2
python-dotenv==0.21.1
